Finished infrastructure and configuration for aws s3 server
