--- .env.dist
@@ -1,21 +1,0 @@
-# NOTE: This port must be above 1024 or else "sudo" is needed when running the app locally
-SERVER_PORT=1111
-
-# Database Details
-DATABASE_PORT=27017
-DATABASE_HOST='127.0.0.1'
-
-# NAF stuff. if we want voice audio and what adapter (socketio or webrtc - need webrtc for audio = true)
-# server url will be 'ws://localhost:8080' for socketio and 
-NAF_AUDIO='false'
-NAF_ADAPTER='socketio'
-NAF_SERVER='/'
-
-# the default password for your default users
-DEFAULT_PASSWORD='password'
-
-#token for JsonwebToken signing
-JWT_SECRET='very_secret'
-
-# Research Settings. Note if you change this you will have to restart server
-ENABLE_RESEARCH_MODE=false
