+++ package.json
@@ -24,6 +24,7 @@
     "jstransformer-markdown-it": "^3.0.0",
     "mongodb": "^6.8.0",
     "mongoose": "^8.5.2",
+    "multer": "^1.4.5-lts.1",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
     "pug": "^3.0.3",
